Organize class files into folders
