+++ helpers/string_parser.rb
